change the key correctly with filter/translate
When translating headers and filtering data from ingredients,
the result ingredient may have different keys comparing to the old one.

Such as, suppose we have a datapoint ingredient with key `country,
year`, when we translate `country` -> `geo`, we should also change
the key to `geo,year`. This is correctly handled in this version.